Breaking out resource table + functionality to its own component. Warning: reorganizes strings.
