+++ kolibri/plugins/coach/assets/src/views/lessons/LessonSummaryPage/index.vue
@@ -87,88 +87,9 @@
         </div>
       </div>
 
-      <core-table>
-        <thead slot="thead">
-          <tr>
-            <th class="visuallyhidden core-table-icon-col">
-              {{ $tr('resourceReorderColumnHeaderForTable') }}
-            </th>
-            <td class="core-table-icon-col">
-              <!-- include header? -->
-            </td>
+      <resource-list-table v-if="lessonResources.length" />
 
-            <th class="core-table-main-col">
-              {{ $tr('nameColumnHeaderForTable') }}
-            </th>
-            <th>
-              {{ $tr('resourceProgressColumnHeaderForTable') }}
-            </th>
-            <th class="visuallyhidden">
-              {{ $tr('resourceRemovalColumnHeaderForTable') }}
-            </th>
-          </tr>
-        </thead>
-        <transition-group
-          name="resource-reorder"
-          slot="tbody"
-          tag="tbody"
-        >
-          <tr
-            v-if="!removals.includes(resourceId)"
-            :key="resourceId"
-            v-for="(resourceId, index) in workingResources"
-          >
-            <!-- TODO add content type icon -->
-            <td class="core-table-icon-col">
-              <ui-icon-button
-                type="flat"
-                icon="keyboard_arrow_up"
-                :ariaLabel="$tr('moveResourceUpButtonDescription')"
-                :disabled="index === 0"
-                @click="moveUpOne(index)"
-                class="position-adjustment-button"
-              />
-              <ui-icon-button
-                type="flat"
-                icon="keyboard_arrow_down"
-                :ariaLabel="$tr('moveResourceDownButtonDescription')"
-                :disabled="index === (workingResources.length - 1)"
-                @click="moveDownOne(index)"
-                class="position-adjustment-button"
-              />
-            </td>
-            <td class="core-table-icon-col">
-              <content-icon :kind="resourceKind(resourceId)" />
-            </td>
-            <td>
-              <k-router-link
-                :to="resourceUserSummaryLink(resourceId)"
-                :text="resourceTitle(resourceId)"
-              />
-            </td>
-            <td>
-              <progress-bar
-                v-if="resourceProgress(resourceId)!==null"
-                class="resource-progress-bar"
-                :progress="resourceProgress(resourceId)"
-                :showPercentage="false"
-              />
-              <span class="progress-message">
-                {{ resourceProgressMessage(resourceId) }}
-              </span>
-            </td>
-            <td>
-              <k-button
-                :text="$tr('resourceRemovalButtonLabel')"
-                @click="stageRemoval(index, resourceId)"
-                appearance="flat-button"
-              />
-            </td>
-          </tr>
-        </transition-group>
-      </core-table>
-
-      <p v-if="workingResources.length===0" class="no-resources-message">
+      <p v-else class="no-resources-message">
         {{ $tr('noResourcesInLesson') }}
       </p>
 
@@ -185,55 +106,33 @@
 
 <script>
 
+  import ResourceListTable from './ResourceListTable';
   import dropdownMenu from 'kolibri.coreVue.components.dropdownMenu';
   import kButton from 'kolibri.coreVue.components.kButton';
-  import uiIconButton from 'keen-ui/src/UiIconButton';
-  import progressBar from 'kolibri.coreVue.components.progressBar';
   import kRouterLink from 'kolibri.coreVue.components.kRouterLink';
-  // why PascalCase?
-  import coreTable from 'kolibri.coreVue.components.CoreTable';
   import map from 'lodash/map';
   import ManageLessonModals from '../ManageLessonModals';
   import { LessonActions, CollectionTypes } from '../../../lessonsConstants';
   import StatusIcon from '../StatusIcon';
   import contentIcon from 'kolibri.coreVue.components.contentIcon';
   import InfoIcon from '../InfoIcon';
-  import { selectionRootLink, resourceUserSummaryLink } from '../lessonsRouterUtils';
-  import { createSnackbar, clearSnackbar } from 'kolibri.coreVue.vuex.actions';
-  import { saveLessonResources } from '../../../state/actions/lessons';
-  import debounce from 'lodash/debounce';
-
-  const removalSnackbarTime = 5000;
-  const saveDebounceTime = 6000;
-
-  // debounced function, pulling out of direct binding to preserve cancel() function
-  const commitRemovals = debounce(function() {
-    const remainingResources = this.workingResources.filter(
-      resourceId => !this.removals.includes(resourceId)
-    );
-    this.removals = [];
-    this.setWorkingResources(remainingResources);
-    this.autoSave();
-  }, removalSnackbarTime);
+  import { selectionRootLink } from '../lessonsRouterUtils';
 
   export default {
     name: 'lessonSummaryPage',
     components: {
+      ResourceListTable,
       dropdownMenu,
       ManageLessonModals,
       StatusIcon,
       contentIcon,
       kButton,
-      uiIconButton,
       kRouterLink,
       InfoIcon,
-      coreTable,
-      progressBar,
     },
     data() {
       return {
         currentAction: null,
-        removals: [],
       };
     },
     computed: {
@@ -256,24 +155,8 @@
       lessonSelectionRootPage() {
         return selectionRootLink({ lessonId: this.lessonId, classId: this.classId });
       },
-      removalMessage() {
-        const numberOfRemovals = this.removals.length;
-
-        if (!numberOfRemovals) {
-          return '';
-        } else if (numberOfRemovals === 1) {
-          return this.$tr('singleResourceRemovalConfirmationMessage', {
-            resourceTitle: this.resourceTitle(this.removals[0]),
-          });
-        }
-
-        return this.$tr('multipleResourceRemovalsConfirmationMessage', {
-          numberOfRemovals,
-        });
-      },
     },
     methods: {
-      resourceUserSummaryLink,
       handleSelectOption({ action }) {
         this.currentAction = action;
       },
@@ -284,87 +167,6 @@
       },
       getGroupName(assignment) {
         return this.learnerGroups.find(lg => lg.id === assignment.collection).name;
-      },
-      resourceTitle(resourceId) {
-        return this.resourceContentNodes[resourceId].title;
-      },
-      resourceKind(resourceId) {
-        return this.resourceContentNodes[resourceId].kind;
-      },
-      resourceProgress(resourceId) {
-        if (this.totalLearners === 0) {
-          return null;
-        }
-        return this.numLearnersCompleted(resourceId) / this.totalLearners;
-      },
-      resourceProgressMessage(resourceId) {
-        return this.$tr('resourceProgressMessage', {
-          completed: this.numLearnersCompleted(resourceId),
-          total: this.totalLearners,
-        });
-      },
-      stageRemoval(index, resourceId) {
-        this.removals.push(resourceId);
-
-        this.createSnackbar({
-          text: this.removalMessage,
-          duration: removalSnackbarTime,
-          autoDismiss: true,
-          actionText: this.$tr('undoActionPrompt'),
-          actionCallback: () => {
-            this.removals = [];
-            this.clearSnackbar();
-          },
-        });
-
-        // cancel any pending calls to reset timer
-        this.cancelCommitRemovals();
-        this.commitRemovals();
-      },
-      commitRemovals,
-      cancelCommitRemovals: commitRemovals.cancel,
-      autoSave() {
-        const modelResources = this.workingResources.map(resourceId => {
-          const node = this.resourceContentNodes[resourceId];
-          return {
-            contentnode_id: node.id,
-            channel_id: node.channel_id,
-            content_id: node.content_id,
-          };
-        });
-        return this.debouncedSaveLessonResources(this.lessonId, modelResources);
-      },
-      moveUpOne(oldIndex) {
-        this.shiftOne(oldIndex, oldIndex - 1);
-      },
-      moveDownOne(oldIndex) {
-        this.shiftOne(oldIndex, oldIndex + 1);
-      },
-      shiftOne(oldIndex, newIndex) {
-        // TODO measure performance, see if worth keeping over generalized shiftMany
-        const resources = [...this.workingResources];
-        const oldResourceId = resources[newIndex];
-        resources[newIndex] = resources[oldIndex];
-        resources[oldIndex] = oldResourceId;
-
-        this.setWorkingResources(resources);
-        this.autoSave();
-
-        this.createSnackbar({
-          text: this.$tr('resourceReorderConfirmationMessage'),
-          autoDismiss: true,
-        });
-      },
-      shiftMany(oldIndex, newIndex) {
-        // to be used w/ drag and drop if we do this
-        const resources = [...this.workingResources];
-        // remove the resourceId from the array, store here
-        const [resourceId] = resources.splice(oldIndex, 1);
-        // re-add resourceId at the new index
-        resources.splice(newIndex, 0, resourceId);
-
-        this.setWorkingResources(resources);
-        this.autoSave();
       },
     },
     vuex: {
@@ -377,31 +179,7 @@
         lessonDescription: state => state.pageState.currentLesson.description,
         lessonAssignments: state => state.pageState.currentLesson.lesson_assignments,
         lessonResources: state => state.pageState.currentLesson.resources,
-        workingResources: state => state.pageState.workingResources,
-        // consider loading this async?
-        resourceContentNodes: state => state.pageState.resourceContentNodes,
         learnerGroups: state => state.pageState.learnerGroups,
-        totalLearners: state => state.pageState.lessonReport.total_learners,
-        numLearnersCompleted(state) {
-          return function counter(contentNodeId) {
-            const report =
-              state.pageState.lessonReport.progress.find(p => p.contentnode_id === contentNodeId) ||
-              {};
-            // If progress couldn't be found, assume 0 learners completed
-            return report.num_learners_completed || 0;
-          };
-        },
-      },
-      actions: {
-        createSnackbar,
-        clearSnackbar,
-        debouncedSaveLessonResources: debounce(saveLessonResources, saveDebounceTime),
-        removeFromWorkingResources(store, resourceId) {
-          store.dispatch('REMOVE_FROM_WORKING_RESOURCES', resourceId);
-        },
-        setWorkingResources(store, resourceArray) {
-          store.dispatch('SET_WORKING_RESOURCES', resourceArray);
-        },
       },
     },
     $trs: {
@@ -425,18 +203,6 @@
       statusTooltipText: 'Active: learners can see lesson. Inactive: hidden from learners.',
       visibleTo: 'Visible to',
       addResourcesButtonPrompt: 'Add resources',
-      resourceReorderColumnHeaderForTable: 'Reorder buttons',
-      nameColumnHeaderForTable: 'Name',
-      resourceProgressColumnHeaderForTable: 'Resource progress',
-      resourceRemovalColumnHeaderForTable: 'Removal button',
-      resourceRemovalButtonLabel: 'Remove',
-      resourceReorderConfirmationMessage: 'New lesson order saved',
-      singleResourceRemovalConfirmationMessage: 'Removed { resourceTitle }',
-      multipleResourceRemovalsConfirmationMessage: 'Removed { numberOfRemovals } resources',
-      moveResourceUpButtonDescription: 'Move this resource one position up in this lesson',
-      moveResourceDownButtonDescription: 'Move this resource one position down in this lesson',
-      undoActionPrompt: 'Undo',
-      resourceProgressMessage: '{completed, number}/{total, number} completed',
     },
   };
 
@@ -516,22 +282,8 @@
       display: table-cell
       text-align: right
 
-  .position-adjustment-button
-    display: block
-    margin: 0
-
-  .resource-reorder-move
-    transition: transform 0.5s
-    background-color: $core-bg-canvas // duping color set in core-table for selected
-
-  .lesson-summary
-    margin-bottom: 30px
-
   .no-resources-message
     text-align: center
     padding: 48px 0
-
-  .progress-message
-    margin-left: 8px
 
 </style>
